Feat: add validations and register users

+++ routes/students.py
@@ -1,8 +1,9 @@
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 from database import get_db
 from models import Student
 from schemas import StudentCreate
+from auth import get_current_user, get_admin_user
 
 router = APIRouter()
 
@@ -11,22 +12,28 @@ def list_students(db: Session = Depends(get_db)):
     return db.query(Student).all()
 
 @router.post("/students")
-def create_student(student: StudentCreate, db: Session = Depends(get_db)):
+def create_student(
+    student: StudentCreate,
+    db: Session = Depends(get_db),
+    admin: dict = Depends(get_admin_user)
+):
     db_student = Student(**student.dict())
     db.add(db_student)
     db.commit()
+    db.refresh(db_student)
     return db_student
 
 @router.put("/students/{student_id}")
 def update_student(
     student_id: int,
     data: StudentCreate,
-    db: Session = Depends(get_db)
+    db: Session = Depends(get_db),
+    admin: dict = Depends(get_admin_user)
 ):
     student = db.query(Student).filter(Student.id == student_id).first()
 
     if not student:
-        return {"error": "Aluno não encontrado"}
+        raise HTTPException(status_code=404, detail="Aluno não encontrado")
 
     student.name = data.name
     student.email = data.email
@@ -37,8 +44,17 @@ def update_student(
     return student
 
 @router.delete("/students/{student_id}")
-def delete_student(student_id: int, db: Session = Depends(get_db)):
+def delete_student(
+    student_id: int,
+    db: Session = Depends(get_db),
+    admin: dict = Depends(get_admin_user)
+):
     db_student = db.query(Student).filter(Student.id == student_id).first()
+
+    if not db_student:
+        raise HTTPException(status_code=404, detail="Aluno não encontrado")
+
     db.delete(db_student)
     db.commit()
+
     return {"message": "Student deleted"}
